Integrating nav mef

+++ Engine/src/Graphics/QueueRenderer.h
@@ -44,7 +44,8 @@ namespace IonixEngine {
 		void Sort();
 		void RenderFromQueue();
 		void ClearQueue(queue<RenderCall>& sprites);
-		void DrawLine(int x1, int y1, int x2, int y2, bool hitColor);
+		void DrawLine(float x1, float y1, float x2, float y2, bool hitColor);
+
 		static QueueRenderer& Get()
 		{
 			static QueueRenderer instance;
